+++ src/main/testgen/StartEndPointData.java
@@ -5,9 +5,9 @@ import algorithms.datatypes.Point;
 public class StartEndPointData {
     public final Point start;
     public final Point end;
-    public final float shortestPath;
+    public final double shortestPath;
     
-    public StartEndPointData(Point start, Point end, float shortestPath) {
+    public StartEndPointData(Point start, Point end, double shortestPath) {
         this.start = start;
         this.end = end;
         this.shortestPath = shortestPath;
